Moved Color Sensor into Feeder
