+++ src/main/java/frc/robot/GameConstants.java
@@ -10,7 +10,6 @@ public class GameConstants {
     public static final int RIGHT_JOYSTICK_ID = 1;
     public static final double DIAG_REL_SPARK_ENCODER = 0.1;
     public static final int DIAG_ABS_SPARK_ENCODER = 20;
-    public static final boolean COLOR_SENSOR_DEBUG = true;
     public static final Color C_PIECE_TARGET = new Color(0.559326171875,0.364013671875,0.0771484375);
     public static final Color C_PLASTIC_TARGET = new Color(0.288330078125,0.478759765625,0.2333984375);
