+++ src/main/java/frc/robot/RobotContainer.java
@@ -28,7 +28,6 @@ import frc.robot.utils.smartshuffleboard.SmartShuffleboard;
 import frc.robot.subsystems.Climber;
 import frc.robot.subsystems.Ramp;
 import frc.robot.commands.StaticClimb;
-import frc.robot.subsystems.ColorSensor;
 
 import java.util.Optional;
 
@@ -47,7 +46,6 @@ public class RobotContainer {
       private final AutoChooser2024 autoChooser;
       private final Shooter shooter = new Shooter();
       private final Feeder feeder = new Feeder();
-      private final ColorSensor colorSensor = new ColorSensor();
       private Climber climber;
 
     /** The container for the robot. Contains subsystems, OI devices, and commands. */
@@ -115,7 +113,7 @@ public class RobotContainer {
         }
         if (Constants.FEEDER_DEBUG){
             SmartShuffleboard.putCommand("Feeder", "Feed", new StartFeeder(feeder));
-            SmartShuffleboard.putCommand("Feeder", "StartFeeder", new FeederColorMatcher(feeder, colorSensor));
+            SmartShuffleboard.putCommand("Feeder", "StartFeeder", new FeederColorMatcher(feeder));
         }
         if (Constants.CLIMBER_DEBUG) {      
           SmartShuffleboard.putCommand("Climber", "Climb", new StaticClimb(climber));
