+++ src/main/java/frc/robot/commands/FeederColorMatcher.java
@@ -6,18 +6,15 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants;
-import frc.robot.subsystems.ColorSensor;
 import frc.robot.subsystems.Feeder;
 import edu.wpi.first.wpilibj.Timer;
 
 public class FeederColorMatcher extends Command {
   private Feeder feeder;
-  private ColorSensor colorSensor;
   private double startTime;
   /** Creates a new FeederColorMatcher. */
-  public FeederColorMatcher(Feeder feeder, ColorSensor colorSensor) {
+  public FeederColorMatcher(Feeder feeder) {
     this.feeder = feeder;
-    this.colorSensor = colorSensor;
     addRequirements(feeder);
   }
 
@@ -43,6 +40,6 @@ public class FeederColorMatcher extends Command {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return (colorSensor.pieceSeen() || Timer.getFPGATimestamp() - startTime > 5.0);
+    return (feeder.pieceSeen() || Timer.getFPGATimestamp() - startTime > 5.0);
   }
 }
